Abstract scheduleList
